requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.3
cssselect2==0.8.0
distro==1.9.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lxml==5.4.0
markdown2==2.5.3
MarkupSafe==3.0.2
narwhals==1.41.0
numpy==2.2.5
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
protobuf==6.31.0
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
PyMuPDF==1.25.5
pypdf==5.5.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.4.1
requests==2.32.3
rpds-py==0.25.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.45.1
streamlit-lottie==0.0.5
svglib==1.5.1
tenacity==9.1.2
tinycss2==1.4.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
watchdog==6.0.0
webencodings==0.5.1
xhtml2pdf==0.2.17



